Added <TwinClassFieldsTable /> to @/widgets
