+++ src/widgets/tables/index.ts
@@ -1,0 +1,1 @@
+export * from "./twinClassFields";
