+++ src/widgets/tables/twinClassFields/index.ts
@@ -1,0 +1,1 @@
+export * from "./twin-class-fields";
